[C] fix: GetV3MaxSpeedMul 파라미터 int→float 수정 + GameSettingsV3.asset 생성 및 연결
- GameSettingsV3.GetV3MaxSpeedMul(int→float): charBaseSpeed가 float 타입이라 CS1503 에러 수정
- Assets/Resources/GameSettingsV3.asset: 신규 생성 (기본값으로 초기화됨)
- GameSettings.asset: v3Settings 슬롯에 GameSettingsV3.asset 연결 (V3 활성화 상태)

+++ Assets/Scripts/Data/GameSettings_V3.cs
@@ -88,7 +88,7 @@ public class GameSettingsV3 : ScriptableObject
     // ──────────────────────────────────────────────────────────────
 
     /// <summary>speed 스탯 → maxSpeedMultiplier (1.0 + speed × perPoint)</summary>
-    public float GetV3MaxSpeedMul(int charSpeed)
+    public float GetV3MaxSpeedMul(float charSpeed)
         => 1f + charSpeed * v3_speedStatPerPoint;
 
     /// <summary>
